The structure of web applications has been completed. Dump and database schema added

+++ src/main/java/com/project/Controller/PointController.java
@@ -3,6 +3,7 @@ package com.project.Controller;
 import com.project.Entity.Point;
 import com.project.Service.PointService;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -17,12 +18,12 @@ public class PointController {
         this.pointService = pointService;
     }
 
-    @GetMapping("/points")
-    public String points(Model model){
-        List<Point> point = pointService.getAll();
-        model.addAttribute("points", point);
-        return "index";
-    }
+//    @GetMapping("/points")
+//    public String points(Model model){
+//        List<Point> point = pointService.getAll();
+//        model.addAttribute("points", point);
+//        return "index";
+//    }
 
     @RequestMapping(path = "/points/new")
     @ResponseStatus(HttpStatus.CREATED)
@@ -30,8 +31,12 @@ public class PointController {
         pointService.save(new Point(name));
     }
     @GetMapping("/points/{id}/delete")
-    @ResponseStatus(HttpStatus.OK)
-    public void deleteTransfer(@PathVariable Long id){
-        pointService.delete(id);
+    public ResponseEntity<?> deleteTransfer(@PathVariable Long id){
+        if(!pointService.getAll().isEmpty() && pointService.getAll().contains(pointService.getById(id))){
+            pointService.delete(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        else
+            return  new ResponseEntity<>(HttpStatus.BAD_GATEWAY);
     }
 }
